Closing of the seisio object
Important operation of closing. If not done, will lead to a problem to
refer to the open file since only a limited amount of references to the
file can be created. That means that after a certain number of
iterations in re-writing the seisio object header, the file may not be
found anymore.