Faster RW method using flipped dimensions
This method is using for loops in Java (in SeisioSDC) to read and write
datasets. The use of flipped dimensions between Matlab and JavaSeis
avoids any permutation calculation time. Now, a multidimensional array
of dimensions (x,y,z) in Matlab corresponds to a multidimensional
object of dimensions (z,y,x) in JavaSeis. The files HeaderRead and
HeaderWrite have been re-written to match this new definition.