Writing tests report
May be the last commit on this branch.

A few tests have been done to show different possible way to write data
using JavaSeis. The most efficient way may be somewhere between doing
all the 'for' loops in Matlab and doing all the 'for' loops in JavaSeis
with a whole copy of the dataset.

2 files should be kept from this branch and replaced in 'JavaSeis'
branch:
- HeaderRead
- HeaderWrite
Those files enable to write a SDC header into a JavaSeis file with a
'flipped dimensions" strategy (the dimensions in JavaSeis are chosen to
be flipped compared to Matlab dimensions). This avoids to do any
permutation as it is currently the case in 'JavaSeis' branch (cf.
FileWrite and HeaderWrite/HeaderRead files in branch 'JavaSeis').